+++ packages/uniforms-material/src/DateField.js
@@ -6,9 +6,10 @@ import React          from 'react';
 import connectField   from 'uniforms/connectField';
 import filterDOMProps from 'uniforms/filterDOMProps';
 
+const DateConstructor = (typeof global === 'object' ? global : window).Date;
 const dateFormat = value => value && value.toISOString().slice(0, -8);
 const dateParse = (timestamp, onChange) => {
-    const date = new Date(timestamp);
+    const date = new DateConstructor(timestamp);
     if (date.getFullYear() < 10000) {
         onChange(date);
     } else if (isNaN(timestamp)) {
@@ -16,7 +17,7 @@ const dateParse = (timestamp, onChange) => {
     }
 };
 
-const Date_ = ({
+const Date = ({
     disabled,
     error,
     errorMessage,
@@ -58,11 +59,9 @@ const Date_ = ({
     </FormControl>
 );
 
-Date_.defaultProps = {
+Date.defaultProps = {
     fullWidth: true,
     margin: 'normal'
 };
 
-Date_.displayName = 'Date';
-
-export default connectField(Date_);
+export default connectField(Date);