+++ packages/uniforms-unstyled/src/DateField.js
@@ -2,9 +2,10 @@ import React          from 'react';
 import connectField   from 'uniforms/connectField';
 import filterDOMProps from 'uniforms/filterDOMProps';
 
+const DateConstructor = (typeof global === 'object' ? global : window).Date;
 const dateFormat = value => value && value.toISOString().slice(0, -8);
 const dateParse = (timestamp, onChange) => {
-    const date = new Date(timestamp);
+    const date = new DateConstructor(timestamp);
     if (date.getFullYear() < 10000) {
         onChange(date);
     } else if (isNaN(timestamp)) {
@@ -12,7 +13,7 @@ const dateParse = (timestamp, onChange) => {
     }
 };
 
-const Date_ = ({
+const Date = ({
     disabled,
     id,
     inputRef,
@@ -47,6 +48,4 @@ const Date_ = ({
     </div>
 ;
 
-Date_.displayName = 'Date';
-
-export default connectField(Date_);
+export default connectField(Date);